logaut user
