+++ src/redux/auth/operationLogin.js
@@ -15,11 +15,6 @@ axios.defaults.baseURL = "https://innov8ors-backend.onrender.com";
 
 axios.defaults.withCredentials = true;
 
-// функція скидання токена авторизації при логауті 
-const clearAuthHeader = () => {
-  axios.defaults.headers.common.Authorization = "";
-};
-
 
 export const login = createAsyncThunk(
   "auth/login",
@@ -64,12 +59,3 @@ export const refreshAccessToken = createAsyncThunk(
   }
 );
 
-//logout юзера
-export const logout = createAsyncThunk("auth/logout", async (_, thunkAPI) => {
-  try {
-     await axios.post("/auth/logout");
-     clearAuthHeader();
-  } catch (error) {
-     return thunkAPI.rejectWithValue(error.message);
-  }
-});
